Se hizo la documentacion de los archivos y se corrigieron datos

+++ Alu.h
@@ -5,7 +5,7 @@
 * \param R1 segundo registro
 * \param R2 tercer registro
 */
-void ADD(unsigned long int *R0,unsigned long int R1,unsigned long int R2);
+void ADD(uint32_t *R0,uint32_t R1,uint32_t R2);
 
 /**
 * \brief funcion ADD que realiza multiplicaciones
@@ -13,7 +13,7 @@ void ADD(unsigned long int *R0,unsigned long int R1,unsigned long int R2);
 * \param R1 segundo registro
 * \param R2 tercer registro
 */
-void AND(unsigned long int *R0,unsigned long int R1,unsigned long int R2);
+void AND(uint32_t *R0,uint32_t R1,uint32_t R2);
 
 /**
 * \brief funcion EOR que reliza la operacion logica XOR
@@ -21,14 +21,14 @@ void AND(unsigned long int *R0,unsigned long int R1,unsigned long int R2);
 * \param R1 segundo registro
 * \param R2 tercer registro
 */
-void EOR(unsigned long int *R0,unsigned long int R1,unsigned long int R2);
+void EOR(uint32_t *R0,uint32_t R1,uint32_t R2);
 
 /**
 * \brief funcion MOV mueve datos de un registro a otro
 * \param *R0 puntero en el que se guarda el dato	
 * \param R1 registro del que se extrae el dato
 */
-void MOV(unsigned long int *R0,unsigned long int R1);
+void MOV(uint32_t *R0,uint32_t R1);
 
 /**
 * \brief funcion EOR que reliza la operacion logica XOR
@@ -36,7 +36,7 @@ void MOV(unsigned long int *R0,unsigned long int R1);
 * \param R1 segundo registro
 * \param R2 tercer registro
 */
-void ORR(unsigned long int *R0,unsigned long int R1,unsigned long int R2);
+void ORR(uint32_t *R0,uint32_t R1,uint32_t R2);
 
 /**
 * \brief funcion sub que realiza restas
@@ -44,4 +44,4 @@ void ORR(unsigned long int *R0,unsigned long int R1,unsigned long int R2);
 * \param R1 segundo registro
 * \param R2 tercer registro
 */
-void SUB(unsigned long int *R0,unsigned long int R1,unsigned long int R2);
+void SUB(uint32_t *R0,uint32_t R1,uint32_t R2);
