+++ banderas.h
@@ -1,1 +1,7 @@
-void banderas(unsigned int *R,unsigned long int R0,unsigned long int R1);
+/**
+* \brief funcion para crear banderas
+* \param *R registro donde se almacenan los resultados
+* \param R0 primer valor a operar
+* \param R1 segundo valor a operar
+*/ 
+void banderas(uint32_t *R,uint32_t R0,uint32_t R1);
