+++ mostrar.h
@@ -2,4 +2,4 @@
 * \brief funcion que muestra los datos del registro
 * \param registros imprime los valores de los registros 
 */
-void mostrardatos(unsigned long int registros[]);
+void mostrardatos(uint32_t registros[]);
